requirements.txt
streamlit
reportlab
streamlit-antd-components
PyPDF2
google-generativeai
streamlit-extras
pymupdf4llm
streamlit-elements
